feat(app and bag): sharing data through context api
