+++ src/pages/bag/Bag.tsx
@@ -1,25 +1,16 @@
-import React, { useEffect, useState } from 'react'
+import React, { useContext } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { getBag } from './service'
 import { ProductPaper, BagDetails } from '../../components/molecules'
 import { Button } from '../../components/atoms'
+import { BagContext } from '../../contexts/BagContext'
 
 const Bag: React.FC = () => {
-  const [bag, setBag] = useState<ProductBag>()
   const navigate = useNavigate()
+  const bag = useContext(BagContext)
 
   const handleGoToPaymentStep = () => {
     navigate('/payment')
   }
-
-  useEffect(() => {
-    const loadBag = async () => {
-      const productsList = await getBag()
-      setBag(productsList)
-    }
-
-    loadBag()
-  }, [])
 
   if (!bag) {
     return <></>
